+++ NuciCraft.API/Service/Models/Coordinates.cs
@@ -4,10 +4,10 @@ namespace NuciCraft.API.Service.Models
     {
         public string World { get; set; }
 
-        public int X { get; set; }
+        public float X { get; set; }
 
-        public int Y { get; set; }
+        public float Y { get; set; }
 
-        public int Z { get; set; }
+        public float Z { get; set; }
     }
 }
